pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "flask-wtf>=1.2.2",
    "boto3>=1.37.34",
    "python-jose>=3.4.0",
    "warrant>=0.6.1",
    "cryptography>=44.0.2",
    "jose>=1.0.0",
    "wtforms>=3.2.1",
    "openai>=1.75.0",
    "sqlalchemy>=2.0.40",
    "botocore>=1.37.34",
    "sendgrid>=6.11.0",
    "pdfplumber>=0.11.6",
    "python-docx>=1.1.2",
    "trafilatura>=2.0.0",
    "twilio>=9.5.2",
    "pandas>=2.2.3",
    "numpy>=2.2.4",
    "python-dateutil>=2.9.0.post0",
    "python-pptx>=1.0.2",
    "beautifulsoup4>=4.13.4",
    "pdfkit>=1.0.0",
    "openpyxl>=3.1.5",
    "markupsafe>=3.0.2",
    "werkzeug>=3.1.3",
    "redis>=5.2.1",
    "elasticsearch>=9.0.0",
    "pika>=1.3.2",
    "anthropic>=0.49.0",
    "pynamodb>=6.0.2",
]
